Add gesture support
This is heavily inspired by what GTK does and lots of the gesture-*
files were copied and modified to work with phoc. The copyright headers
were modified to mention the original LGPL-2.1 license and authors.

It allows us to track drag and swipe gestures with mouse and touch and
should be extensible to other gestures easily.